+++ artemis-fluid/artemis-fluid-maven-integration-test/artemis-fluid-maven-integration-test-components/src/main/java/com/artemis/ProxyExperiment.java
@@ -1,0 +1,8 @@
+package com.artemis;
+
+/**
+ * @author Daan van Yperen
+ */
+public class ProxyExperiment {
+    public int x, y;
+}